flatten heirachy
