--- tests/regression/__init__.py
@@ -1,1 +1,0 @@
-# this file marks the directory as a python package
