+++ tests/requirements.txt
@@ -2,3 +2,4 @@ pytest
 cogent3
 black
 pyYAML
+numpy
